requirements.txt
# Backend IA
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.24.0

# Frontend Web
flask>=3.0.0
requests>=2.31.0

# Traitement d'images
pillow>=10.0.0
python-multipart>=0.0.6

# Utilitaires
python-dotenv>=1.0.0
huggingface-hub>=0.18.0
